Update Portfolio
made significant updates to the visual layout of the "popup" form, and began an overhaul of the JS. many changes to the JS are yet to come, and most of order66() is not functional in its current state. This commit is basically just to back up the 20-or-so hours I've spent so far updating everything.

Changes to come:
--Probable update to coloring
--making order66() fully functional again